Move schema generation functionality into phdi_building_blocks.schemas and add required dependencies to with poetry.
