+++ FHIR-parquet-explore/requirements.txt
@@ -4,4 +4,4 @@ pyarrow
 pandas
 numpy
 pandas
-phdi_building_blocks @ git+https://github.com/CDCgov/prime-public-health-data-infrastructure@main#subdirectory=src/lib/phdi-building-blocks
+phdi_building_blocks @ git+https://github.com/CDCgov/prime-public-health-data-infrastructure@dan/initial-fhir-to-parquet#subdirectory=src/lib/phdi-building-blocks
