+++ src/lib/phdi-building-blocks/pyproject.toml
@@ -14,6 +14,9 @@ azure-storage-blob = "^12.12.0"
 polling = "^0.3.2"
 phonenumbers = "^8.12.48"
 pycountry = "^22.3.5"
+PyYAML = "^6.0"
+pyarrow = "^8.0.0"
+pandas = "^1.4.2"
 
 [tool.poetry.dev-dependencies]
 Sphinx = "^4.4.0"
